Update content-type, move sqlite into backend folder
