+++ Makefile
@@ -3,7 +3,7 @@
 BUILD_TAGS = -tags json1
 BUILD_PATH = build/cabby
 CLI_FILES = $(shell find cmd/cabby-cli/*.go -name '*go' | grep -v test)
-PACKAGES = ./ sqlite http cmd/cabby-cli
+PACKAGES = ./ backends/sqlite http cmd/cabby-cli
 
 all: config/cabby.json cert dependencies dev-db
