+++ backends/sqlite/api_root_test.go
