+++ backends/sqlite/collection.go
