+++ backends/sqlite/discovery.go
