+++ backends/sqlite/discovery_test.go
