+++ backends/sqlite/manifest.go
