+++ backends/sqlite/manifest_test.go
