+++ backends/sqlite/object_test.go
