+++ backends/sqlite/testdata/bundle_invalid_objects.json
