+++ backends/sqlite/testdata/malware_bundle.json
