+++ backends/sqlite/user.go
