+++ backends/sqlite/user_test.go
